gjort getrecipes till en js fil
